+++ changelog.md
@@ -7,6 +7,7 @@
 - Adding color palette options for the customizer.
 - Added Distraction free checkout option.
 - Added 2 steps checkout option.
+- Added in the default banner background colour and title support for the LSX Blocks default title.
 
 ### Updated